Remove TS-Node
