+++ package.json
@@ -4,21 +4,12 @@
     "private": true,
     "main": "dist/index.js",
     "scripts": {
-        "start": "ts-node src/index",
+        "build": "tsc",
+        "start": "node build/index.js",
         "dev": "ts-node-dev src/index.ts",
         "test": "jest --watch"
     },
     "dependencies": {
-        "@types/bcryptjs": "^2.4.2",
-        "@types/cors": "^2.8.6",
-        "@types/dotenv": "^8.2.0",
-        "@types/express": "^4.17.2",
-        "@types/express-session": "^1.15.15",
-        "@types/helmet": "^0.0.45",
-        "@types/jsonwebtoken": "^8.3.5",
-        "@types/node": "^13.1.4",
-        "@types/pg": "^7.11.2",
-        "@types/supertest": "^2.0.8",
         "bcryptjs": "^2.4.3",
         "class-transformer": "^0.2.3",
         "class-transformer-validator": "^0.8.0",
@@ -34,15 +25,28 @@
         "secure-random-password": "^0.2.1",
         "segal-decorators": "^0.3.1",
         "stripe": "^8.4.0",
-        "ts-node": "^8.5.4",
-        "typeorm": "^0.2.21",
-        "typescript": "^3.7.2"
+        "typeorm": "^0.2.21"
     },
     "devDependencies": {
+        "@types/bcryptjs": "^2.4.2",
+        "@types/cors": "^2.8.6",
+        "@types/dotenv": "^8.2.0",
+        "@types/express": "^4.17.2",
+        "@types/express-session": "^1.15.15",
+        "@types/helmet": "^0.0.45",
         "@types/jest": "^24.0.24",
+        "@types/jsonwebtoken": "^8.3.5",
+        "@types/node": "^13.1.4",
+        "@types/pg": "^7.11.2",
+        "@types/supertest": "^2.0.8",
+        "@types/webpack": "^4.41.3",
         "jest": "^24.9.0",
         "supertest": "^4.0.2",
         "ts-jest": "^24.2.0",
-        "ts-node-dev": "^1.0.0-pre.44"
+        "ts-loader": "^6.2.1",
+        "ts-node-dev": "^1.0.0-pre.44",
+        "typescript": "^3.7.2",
+        "webpack": "^4.41.5",
+        "webpack-cli": "^3.3.10"
     }
 }
