+++ tsconfig.json
@@ -3,7 +3,6 @@
         "emitDecoratorMetadata": true,
         "experimentalDecorators": true,
         "target": "es5",
-        "sourceMap": true,
         "outDir": "./build/"
     }
 }
